identity_service/pyproject.toml
[tool.poetry]
name = "identity_service"
version = "0.1.0"
description = ""
authors = ["Nhsdkk <38581029+Nhsdkk@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.62.1"
sqlalchemy = "^2.0.28"
jwcrypto = "^1.5.6"
grpcio-tools = "^1.62.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pre-commit = "^3.5.0"
flake8-docstrings = "^1.7.0"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
